Correct misleading error message
The old message was misleading because it implied that the change which
is about to be pushed into review is already on the upstream branch.
But this is not the case since the change can be on any other upstream
branch.

For example:
    * test commit [branch][origin/branch]
    |
    * commit [master][origin/master]

    git review master would yield:
    No changes between HEAD and origin/master. Submitting for review
    would be pointless.

Right now I don't have a better idea how to tweak the "git log" command
to yield exactly the information needed here.